Add sender and receiver modules and scheme configuration option.
